Fixed TestLauncher/LauncherTest
