+++ tests/gfx/LauncherTest.png
